Redesign login form and improve code documentation
This commit includes a complete redesign of the login form in `frmDangNhap.Designer.cs`, replacing the `GroupBox` with a `Panel` that features minimize and close buttons. The layout and appearance of controls have been enhanced, including a flat style for the `buttonLogin`.

In `frmDangNhap.cs`, the constructor now sets the `FlatStyle` of the `buttonLogin` and introduces a new `frmDangNhap_Load` event handler.

Additionally, a comment has been added in `frmListOrder.cs` to clarify the invoice printing process, improving code documentation.